Workaround @flow pragma issue
